CodeGen.java
Bootstrap Code Generator
CodeGen.java
codegen.xml

A little gift to all of you: what I call creative laziness.

This code generator is just a beginning. It has the shell. It must be completed for the different options and modules available in bootstrap. However, when you do complete it, it will save you so much time.